refactor[server]: update package structure
