+++ server/src/main/java/com/github/wizard/game/Color.java
@@ -1,0 +1,8 @@
+package com.github.wizard.game;
+
+public enum Color {
+    RED,
+    GREEN,
+    BLUE,
+    YELLOW,
+}
